+++ Source/Core/Color.js
@@ -633,6 +633,27 @@ Color.prototype.toCssColorString = function () {
 };
 
 /**
+ * Creates a string containing CSS hex string color value for this color.
+ *
+ * @returns {String} The CSS hex string equivalent of this color.
+ */
+Color.prototype.toCssHexString = function () {
+  var r = Color.floatToByte(this.red).toString(16);
+  if (r.length < 2) {
+    r = "0" + r;
+  }
+  var g = Color.floatToByte(this.green).toString(16);
+  if (g.length < 2) {
+    g = "0" + g;
+  }
+  var b = Color.floatToByte(this.blue).toString(16);
+  if (b.length < 2) {
+    b = "0" + b;
+  }
+  return "#" + r + g + b;
+};
+
+/**
  * Converts this color to an array of red, green, blue, and alpha values
  * that are in the range of 0 to 255.
  *